moved comment methods to comment classes, changed path for endpoints
