+++ backend/post-service/src/main/java/com/speakapp/postservice/controllers/CommentController.java
@@ -1,5 +1,7 @@
 package com.speakapp.postservice.controllers;
 
+import com.speakapp.postservice.dtos.CommentCreateDTO;
+import com.speakapp.postservice.dtos.CommentGetDTO;
 import com.speakapp.postservice.dtos.CommentPageGetDTO;
 import com.speakapp.postservice.services.CommentService;
 import lombok.RequiredArgsConstructor;
@@ -22,5 +24,12 @@ public class CommentController {
                                                 @RequestParam UUID postId,
                                                 @RequestHeader("UserId") UUID userId ){
         return commentService.getCommentsForPost(pageNumber, pageSize, postId, userId);
+    }
+
+
+    @PostMapping("")
+    @ResponseStatus(HttpStatus.CREATED)
+    public CommentGetDTO createComment(@RequestBody CommentCreateDTO commentCreateDTO, @RequestHeader("UserId") UUID userId) {
+        return commentService.createComment(commentCreateDTO, userId);
     }
 }
