+++ backend/post-service/src/main/java/com/speakapp/postservice/services/PostService.java
@@ -46,32 +46,6 @@ public class PostService {
         );
     }
 
-    public CommentGetDTO createComment(CommentCreateDTO commentCreateDTO, UUID userId) {
-
-        Post postToBeCommented = postRepository.findById(commentCreateDTO.getPostId()).orElseThrow(()->
-        new ResponseStatusException(HttpStatus.NOT_FOUND, "Post with id = " + commentCreateDTO.getPostId() + " was not found"));
-
-        int lengthOfContent = commentCreateDTO.getContent().length();
-
-        if(lengthOfContent > 500 || lengthOfContent == 0){
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Your comment can't be empty and can have maximally 500 characters");
-        }
-
-        UserGetDTO author = userServiceCommunicationClient.getUserById(userId);
-
-        Comment savedComment = commentRepository.save(commentMapper.toEntity(commentCreateDTO.getContent(),
-                postToBeCommented,
-                userId));
-
-        ReactionsGetDTO reactionsGetDTO = reactionsMapper.toGetDTO(Collections.emptyMap());
-
-        return commentMapper.toGetDTO(savedComment,
-                author,
-                reactionsGetDTO,
-                null);
-
-    }
-
     public PostGetDTO updatePost(PostCreateDTO postCreateDTO, UUID postId, UUID userId) {
         Post postToUpdate = postRepository.findById(postId).orElseThrow(() ->
                 new ResponseStatusException(HttpStatus.NOT_FOUND, "Post with id = " + postId + " was not found"));
